Removed the FIXME TODO build rule.
Basically, we were getting false positives from Realm Swift.... And I
could figure out how to make this rule, ignore the Carthage folder...
There is a swiftlint rule which checks for FIXME and TODO's already.

This rule was a bit more flexible in that it looked for
```
KEYWORDS="TODO|FIXME|TEMP|\?\?\?:|\!\!\!:"
```
but No one is yet addind `//???` or `//!!!` so let's just not do that
going forward.