Be able to log without a message.
There were a few places where we logged the file and function, without
really passing a message.
This was a bit silly and redundant because the file and function are
logged by default with all messages. It is nice for tracing purposes, to
be able to see these simple flow log messages "This function got called"
so by default we'll just log a dash no need for you to put anything in
the message if you don't want to.
The default message will look like this now

```
2017-07-06 17:08:08.302 WorkWeek[57189:1841438] [debug] (start(), AppCoordinator.swift:20) : -
```