Now there is only one true location manager.....
AppDelegate -> LocationManager
This reference is passed around to all the other objects which need it.

It feels a tiny bit not quite right yet, but we're working on it.
I'd rather not just use a singleton instance. I like the idea of all the
objects, being setup / configured with the helpers they need, instead of
the helpers they need "just magically/hopefully" existing out in the
world.

But we might change it because this seems a little bit like too much...
we'll see