Add basic search.
When a search is performed the search for your query is logged, and the
search is performed by apple's service, on the current visible map
region.

You probably noticed my `Result`... basically when an API gives you, a
completion handler of type `(thing: Thing?, error: Error?) -> Void`
there is no guarantee that both of those things are not `nil` so maybe
you perform a search but you get back a nil response and a nil error! We
now have a type system that can help us deal with these either or
situations. So we'll encode it.