Remove the address label, it never got hooked up
I'm happy to re-add it later if we think it provides value. I'd feel bad
shipping a beta to some folks I barely know, with that dumb thing just
chilling there. Better to re-add it later if we think it provides value,
but right now it's just messing up our beautiful interface